refactor(web): reset share dialog state via mount, not an effect
Split the link body into ShareDialogLinks, which the dialog mounts only
while open — so its link + copied state starts fresh on each open instead
of being reset through a prop-driven effect.
